\name{RcppKmeans}
\alias{Kmeans}
\docType{package}

\title{RcppKmeans}

\description{
    Rcpp Kmeans
}


\author{
颜烈彬 <yanlb@ucweb.com>
}

\usage{
    Kmeans(strings, cluster.no=10, max.iter=100, epsilon=0.001)
}

\arguments{
    \item{strings} is a list of string, which a string is a list of words.
}
\details{%
    The measure used in our implement is cosine distance.
}
\references{

}
\keyword{ kmeans }
